whole lot of stuff

+++ libft/ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: sonfong <sonfong@student.42kl.edu.my>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/27 15:41:02 by sonfong           #+#    #+#             */
-/*   Updated: 2026/07/28 11:36:44 by sonfong          ###   ########.fr       */
+/*   Updated: 2026/08/05 18:38:38 by sonfong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
